+++ CHANGELOG.md
@@ -4,6 +4,8 @@ FEATURES:
 
   - Rename BROWSER_ env variables conventions to OJS_
   - Add configurable port and ip binding of node process through OJS_NODEJS_PORT and OJS_NODEJS_IP env variables
+  - Add url property on Request class, based on Url class (https://url.spec.whatwg.org/)
+  - Refactory RouteUtils.reverse() signature adding request object as required
 
 BUG FIXES: